Updated Feature Section anchor links
jump links actually work now, fixed by adding: 
            <!-- Anchor Add-On -->
            <div style="visibility: hidden; width: 0; height: 0; position: relative;">
              <div id="featured-project1" style="visibility: hidden; width: 0; height: 0; position: absolute; top: -300px;"></div>
            </div> 
onto the containers/boxes for each of the corresponding boxes. Adding it as a separate div outside of the container/boxes of the projects would result in items shifting as it would take up space despite being invisible and size set to zero.